requirements.txt
streamlit
pytesseract
pdf2image
Pillow
spacy
azure-storage-blob
